Moving Linux runtime/plugin components out of executable path
